# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/ferenc/pamp/data/model/home/good_deal/GoodDealResponse.java
#	app/src/main/java/com/ferenc/pamp/presentation/screens/main/chat/messenger/MessengerContract.java
#	app/src/main/java/com/ferenc/pamp/presentation/screens/main/chat/messenger/MessengerPresenter.java
#	app/src/main/java/com/ferenc/pamp/presentation/screens/main/chat/orders/OrderContract.java
#	app/src/main/java/com/ferenc/pamp/presentation/service/NotificationsService.java
#	app/src/main/res/layout/fragment_chat_orders.xml

+++ app/src/main/java/com/ferenc/pamp/domain/SocketRepository.java
@@ -106,9 +106,9 @@ public class SocketRepository implements MessengerContract.SocketModel {
 
                 description.quantity = data.has(valDescription) ? (float) (data.getJSONObject(valDescription).has(valQuantity) ? data.getJSONObject(valDescription).getDouble(valQuantity) : 0) : 0;
                 description.firstName = (data.has(valDescription) ? data.getJSONObject(valDescription).has(valFirstName) ? data.getJSONObject(valDescription).getString(valFirstName) : "" : "");
-                description.deliveryEndDate = data.has(valDescription) ? data.getJSONObject(valDescription).has(valDeliveryEndDate) ? data.getJSONObject(valDescription).getInt(valDeliveryEndDate) : 0 : 0;
-                description.deliveryStartDate = data.has(valDescription) ? data.getJSONObject(valDescription).has(valDeliveryStartDate) ? data.getJSONObject(valDescription).getInt(valDeliveryStartDate) : 0 : 0;
-                description.closingDate = data.has(valDescription) ? data.getJSONObject(valDescription).has(valClosingDate) ? data.getJSONObject(valDescription).getInt(valClosingDate) : 0 : 0;
+                description.deliveryEndDate = data.has(valDescription) ? data.getJSONObject(valDescription).has(valDeliveryEndDate) ? data.getJSONObject(valDescription).getLong(valDeliveryEndDate) : 0 : 0;
+                description.deliveryStartDate = data.has(valDescription) ? data.getJSONObject(valDescription).has(valDeliveryStartDate) ? data.getJSONObject(valDescription).getLong(valDeliveryStartDate) : 0 : 0;
+                description.closingDate = data.has(valDescription) ? data.getJSONObject(valDescription).has(valClosingDate) ? data.getJSONObject(valDescription).getLong(valClosingDate) : 0 : 0;
 
                 messageResponse.description = description;
